+++ jslib/mappanel/Changelog.md
@@ -1,3 +1,11 @@
+* v0.0.4 2012-12-05T14:22:43 UTC+0800
+  - New Features
+    * Supported `command + enter`
+
+* v0.0.3 2012-12-05T14:22:43 UTC+0800
+  - New Features
+    * Big Map
+
 * v0.0.2 2012-11-27T13:38:30 UTC+0800
   * try-catch `google map`